backend/requirements.txt
requests==2.28.2
uvicorn==0.21.1
fastapi==0.95.0
fastapi-jwt-auth[asymmetric]
fastapi_mail==1.2.7
jinja2
fastapi-cache2[redis]
joblib==1.2.0
loguru==0.6.0
starlette~=0.26.1
pydantic~=1.10.7
pydantic[dotenv, email]
dadata==21.10.1
alembic==1.10.2
greenlet==2.0.2
protobuf==4.22.1
python-jose==3.3.0
passlib==1.7.4
cryptography==40.0.1
sqlalchemy==1.4.47
asyncpg==0.27.0
psycopg2-binary==2.9.5
python-multipart
aioboto3
pillow==9.4.0
aiofiles
websockets==10.4


python-engineio==3.14.2
python-socketio==4.6.1
